+++ global_helpers/panther_crowdstrike_fdr_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_crowdstrike_fdr_helpers.py
+GlobalID: "panther_crowdstrike_fdr_helpers"
+Description: >
+  Used to define global helpers and variables for CrowdStrike FDR events.
