+++ global_helpers/panther_github_helpers.py
@@ -1,0 +1,9 @@
+def github_alert_context(event):
+    return {
+        "action": event.get("action", ""),
+        "actor": event.get("actor", ""),
+        "actor_location": event.deep_get("actor_location", "country_code"),
+        "org": event.get("org", ""),
+        "repo": event.get("repo", ""),
+        "user": event.get("user", ""),
+    }
