+++ global_helpers/panther_github_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_github_helpers.py
+GlobalID: "panther_github_helpers"
+Description: >
+  Used to define global helpers and variables for GitHub events.
