+++ global_helpers/panther_slack_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_slack_helpers.py
+GlobalID: "panther_slack_helpers"
+Description: >
+  Used to define global helpers and variables for Slack events.
