+++ rules/crowdstrike_rules/crowdstrike_detection_passthrough.py
@@ -1,4 +1,7 @@
-from panther_base_helpers import crowdstrike_detection_alert_context, get_crowdstrike_field
+from panther_crowdstrike_fdr_helpers import (
+    crowdstrike_detection_alert_context,
+    get_crowdstrike_field,
+)
 
 
 def rule(event):
