+++ rules/crowdstrike_rules/crowdstrike_macos_add_trusted_cert.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import crowdstrike_process_alert_context
+from panther_crowdstrike_fdr_helpers import crowdstrike_process_alert_context
 
 
 def rule(event):
