+++ rules/crowdstrike_rules/crowdstrike_remote_access_tool_execution.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import crowdstrike_detection_alert_context
+from panther_crowdstrike_fdr_helpers import crowdstrike_detection_alert_context
 
 REMOTE_ACCESS_EXECUTABLES = {
     "teamviewer_service.exe",
