+++ rules/crowdstrike_rules/crowdstrike_systemlog_tampering.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import crowdstrike_detection_alert_context
+from panther_crowdstrike_fdr_helpers import crowdstrike_detection_alert_context
 
 CLEARING_SYSTEM_LOG_TOOLS = {
     "wevtutil.exe": ["cl", "clear-log"],
