+++ rules/crowdstrike_rules/crowdstrike_unusual_parent_child_processes.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import crowdstrike_detection_alert_context
+from panther_crowdstrike_fdr_helpers import crowdstrike_detection_alert_context
 
 SUSPICIOUS_PARENT_CHILD_COMBINATIONS_WINDOWS = {
     ("svchost.exe", "cmd.exe"),
