+++ rules/github_rules/github_repo_vulnerability_dismissed.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import github_alert_context
+from panther_github_helpers import github_alert_context
 
 
 def rule(event):
