+++ rules/slack_rules/slack_app_removed.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import slack_alert_context
+from panther_slack_helpers import slack_alert_context
 
 APP_REMOVED_ACTIONS = [
     "app_restricted",
