+++ rules/slack_rules/slack_private_channel_made_public.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import slack_alert_context
+from panther_slack_helpers import slack_alert_context
 
 
 def rule(event):
